safari-600.7-branch/LayoutTests/http/tests/navigation/target-blank-opener-post-expected.txt

CONSOLE MESSAGE: line 3: Referrer: http://127.0.0.1:8000/navigation/target-blank-opener-post.html
CONSOLE MESSAGE: line 7: window.opener: PASS
Test that rel="noreferrer" is unsupported by form elements, and that window.opener is available when form submission opens a new window.
